dish and restaurandId statistics fix

+++ src/Delivery.Application/Dtos/StatisticDtos/DishRevenueStatisticsResponse.cs
@@ -1,7 +1,7 @@
 ﻿public class DishRevenueStatisticsResponse
 {
     public List<DailyDishRevenue> Daily { get; set; } = new();
-    public decimal TotalRevenue { get; set; }
+    public double TotalRevenue { get; set; }
     public int TotalOrders { get; set; }
-    public decimal AverageRevenue { get; set; }
+    public double AverageRevenue { get; set; }
 }
