+++ src/Delivery.Application/Dtos/StatisticDtos/RestaurantRevenueStatisticsDto.cs
@@ -1,7 +1,7 @@
 ﻿public class RestaurantRevenueStatisticsDto
 {
     public List<RestaurantDailyRevenueDto> Daily { get; set; }
-    public decimal TotalRevenue { get; set; }
+    public double TotalRevenue { get; set; }
     public int OrderCount { get; set; }
-    public decimal AverageOrderValue { get; set; }
+    public double AverageOrderValue { get; set; }
 }
